requirements.txt
pyautogui
pillow
numpy
pytesseract
opencv-python
sympy
#cnocr
